make async io always return RUNNING on first call, fix clockcache refcount size bug

+++ src/clockcache.h
@@ -17,7 +17,7 @@
 #define TRACE_ADDR  (UINT64_MAX - 1)
 #define TRACE_ENTRY (UINT32_MAX - 1)
 
-//#define RECORD_ACQUISITION_STACKS
+// #define RECORD_ACQUISITION_STACKS
 
 /* how distributed the rw locks are */
 #define CC_RC_WIDTH 4
@@ -123,9 +123,9 @@ struct clockcache {
    platform_heap_id     heap_id;
 
    // Distributed locks (the write bit is in the status uint32 of the entry)
-   buffer_handle   rc_bh;
-   volatile uint8 *refcount;
-   volatile uint8 *pincount;
+   buffer_handle    rc_bh;
+   volatile uint16 *refcount;
+   volatile uint8  *pincount;
 
    // Clock hands and related metadata
    volatile uint32  evict_hand;
